feat: pipeline scheduling (#7)

+++ src/main/kotlin/com/github/kronenthaler/ghasimulator/stats/Stats.kt
@@ -1,6 +1,6 @@
 package com.github.kronenthaler.ghasimulator
 
-data class PipelineStats(val startTime: Long, val endTime: Long, val totalQueueTime: Long, val jobCount: Long) {
+data class PipelineStats(val startTime: Long, val endTime: Long, val totalQueueTime: Long, val jobCount: Int) {
     fun toString(timescale: Int): String {
         return "${(endTime - startTime) / timescale}\t$startTime\t$endTime\t$totalQueueTime\t$jobCount"
     }
